refactor: 역에 대해서 패키지 정리
