+++ src/main/java/subway/station/exception/StationNotFoundException.java
@@ -1,4 +1,4 @@
-package subway.station;
+package subway.station.exception;
 
 public class StationNotFoundException extends RuntimeException {
