+++ src/main/java/subway/station/repository/StationRepository.java
@@ -1,6 +1,8 @@
-package subway.station;
+package subway.station.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import subway.station.domain.Station;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
 }
