+++ src/main/java/subway/station/service/StationService.java
@@ -1,4 +1,4 @@
-package subway.station;
+package subway.station.service;
 
 import java.util.List;
 import java.util.Optional;
@@ -6,6 +6,11 @@ import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import subway.station.domain.Station;
+import subway.station.repository.StationRepository;
+import subway.station.view.StationRequest;
+import subway.station.view.StationResponse;
 
 @Service
 @Transactional(readOnly = true)
